+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - PR #612 Prevent an exception from occuring with true division on integer series.
 - PR #630 Fix deprecation warning for `pd.core.common.is_categorical_dtype`
 - PR #622 Fix Series.append() behaviour when appending values with different numeric dtype
+- PR #673 Fix array of strings not being caught in from_pandas
 - PR #644 Fix return type and column support of dataframe.quantile()
 - PR #634 Fix create `DataFrame.from_pandas()` with numeric column names
 - PR #654 Add resolution check for GDF_TIMESTAMP in Join